update(framework): update to IntelliJ Platform Gradle Plugin 2.x
