+++ build.gradle.kts
@@ -1,8 +1,26 @@
+import org.gradle.kotlin.dsl.withType
+import org.jetbrains.intellij.platform.gradle.tasks.PublishPluginTask
+import org.jetbrains.intellij.platform.gradle.tasks.RunIdeTask
+
 plugins {
+    id("org.jetbrains.intellij.platform")
 }
 
-allprojects {
-    repositories {
-        mavenCentral()
+val publishPlugin by tasks.existing(PublishPluginTask::class) {
+    token.set(getPropertyValue("publishToken"))
+}
+
+tasks.withType<RunIdeTask> {
+    jvmArgs("-Xms4g", "-Xmx4g", "-Dcom.sun.management.jmxremote")
+}
+
+fun getPropertyValue(name: String): String? {
+    return if (project.extra.has(name)) project.extra[name]?.toString() else System.getenv(name)
+}
+
+dependencies {
+    intellijPlatform {
+        intellijIdeaCommunity("2025.1.2")
+        pluginModule(implementation(project(":move-tab-plugin")))
     }
 }
