+++ settings.gradle.kts
@@ -1,3 +1,5 @@
+import org.jetbrains.intellij.platform.gradle.extensions.intellijPlatform
+
 pluginManagement {
     repositories {
         mavenCentral()
@@ -9,6 +11,7 @@ plugins {
     id("com.gradle.develocity") version "4.0.2"
     id("com.gradle.common-custom-user-data-gradle-plugin") version "2.3"
     id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
+    id("org.jetbrains.intellij.platform.settings") version "2.6.0"
 }
 
 rootProject.name = "intellij-move-tab"
@@ -26,6 +29,16 @@ develocity {
             hostname { "<obfuscated>" }
             ipAddresses { listOf("<obfuscated>") }
             username { "<obfuscated>" }
+        }
+    }
+}
+
+dependencyResolutionManagement {
+    repositoriesMode = RepositoriesMode.FAIL_ON_PROJECT_REPOS
+    repositories {
+        mavenCentral()
+        intellijPlatform {
+            defaultRepositories()
         }
     }
 }
